ci(screener): disable screener PR checks (#5253)
**Related Issue:** NA

## Summary
Unfortunately we need to temporarily disable screener to prevent insane overage fees. We are working on a short term solution to turn screener back on, and a mid-term solution to switch to Chromatic. Screener can still be run on a branch manually following these steps: https://docs.github.com/en/actions/managing-workflow-runs/manually-running-a-workflow

For now, please try not to manually run screener unless absolutely necessary. We only have around 35 runs per month at our current snapshot count. We will be working to reduce the number of stories so we can reenable screener PR checks in the short term.
<!--

Please make sure the PR title and/or commit message adheres to the https://www.conventionalcommits.org/en/v1.0.0/ specification.

Note: If your PR only has one commit and it is NOT semantic, you will need to either

a. add another commit and wait for the check to update
b. proceed to squash merge, but make sure the commit message is the same as the title.

This is because of the way GitHub handles single-commit squash merges (see https://github.com/zeke/semantic-pull-requests/issues/17)

If this is component-related, please verify that:

- [ ] feature or fix has a corresponding test
- [ ] changes have been tested with demo page in Edge

---

If this is skipping an unstable test:

- include info about the test failure
- submit an unstable-test issue by [choosing](https://github.com/Esri/calcite-components/issues/new/choose) the unstable test template and filling it out

-->