ci(screener): clean up and exclude unnecessary stories (#5257)
**Related Issue:** #5253

## Summary
This PR cleanups up and combines stories to lower our snapshot count
below 40k per month.

- Append `_NoTest` to story names that we don't want screener to
snapshot.
- Append `_TestOnly` to story names that we do not want to appear in the
[Storybook deployment on github
pages](https://esri.github.io/calcite-components/?path=/story/overview-home--page).
- Set up filters to accomplish the above two points. The filter
determines which stories to skip using a boolean environment variable:
`STORYBOOK_SCREENSHOT_TEST_BUILD`. The default is `NoTest`, and the env
var is set to `true` in the [`pr-screener`
workflow](https://github.com/Esri/calcite-components/blob/master/.github/workflows/pr-screener.yml).
- Combine stories, e.g. if a component has "DarkTheme" and "RTL"
snapshots, we combined them into a single `darkThemeRTL_TestOnly`
snapshot.
- Consistency cleanups, including camel case story names and using
"simple" as the name for the first story (rather than "basic").

<!--

Please make sure the PR title and/or commit message adheres to the
https://www.conventionalcommits.org/en/v1.0.0/ specification.

Note: If your PR only has one commit and it is NOT semantic, you will
need to either

a. add another commit and wait for the check to update
b. proceed to squash merge, but make sure the commit message is the same
as the title.

This is because of the way GitHub handles single-commit squash merges
(see https://github.com/zeke/semantic-pull-requests/issues/17)

If this is component-related, please verify that:

- [ ] feature or fix has a corresponding test
- [ ] changes have been tested with demo page in Edge

---

If this is skipping an unstable test:

- include info about the test failure
- submit an unstable-test issue by
[choosing](https://github.com/Esri/calcite-components/issues/new/choose)
the unstable test template and filling it out

-->

Co-authored-by: JC Franco <jfranco@esri.com>
Co-authored-by: Erik Harper <eharperdicianno@esri.com>