test(custom-theme): use placeholderImage (#5853)
**Related Issue:** #5812

## Summary
The image service we were using for the card's thumbnail in the custom
theme story is/was down, which caused CI failures. For example:

https://www.chromatic.com/test?appId=6266d45509d7eb004aa254fb&id=6386b5d2e340d812b9c9a134

I switched it out to use our placeholderImage util.
<!--

Please make sure the PR title and/or commit message adheres to the
https://www.conventionalcommits.org/en/v1.0.0/ specification.

Note: If your PR only has one commit and it is NOT semantic, you will
need to either

a. add another commit and wait for the check to update
b. proceed to squash merge, but make sure the commit message is the same
as the title.

This is because of the way GitHub handles single-commit squash merges
(see https://github.com/zeke/semantic-pull-requests/issues/17)

If this is component-related, please verify that:

- [ ] feature or fix has a corresponding test
- [ ] changes have been tested with demo page in Edge

---

If this is skipping an unstable test:

- include info about the test failure
- submit an unstable-test issue by
[choosing](https://github.com/Esri/calcite-components/issues/new/choose)
the unstable test template and filling it out

-->