build(tsconfig): bump target to es2020 and lib to es2021 (#5879)
**Related Issue:** #

## Summary

Creating this PR based on the Teams discussion about increasing our
`target`. I confirmed with @dasa that these changes are safe for our
stakeholders. But any `target` higher than 2020 has the potential to
cause them problems.

lib 2021 has some useful stuff we can benefit from, like `Promise.any`,
`string.prototype.replaceAll`, and logical assignment operators.

<!--

Please make sure the PR title and/or commit message adheres to the
https://www.conventionalcommits.org/en/v1.0.0/ specification.

Note: If your PR only has one commit and it is NOT semantic, you will
need to either

a. add another commit and wait for the check to update
b. proceed to squash merge, but make sure the commit message is the same
as the title.

This is because of the way GitHub handles single-commit squash merges
(see https://github.com/zeke/semantic-pull-requests/issues/17)

If this is component-related, please verify that:

- [ ] feature or fix has a corresponding test
- [ ] changes have been tested with demo page in Edge

---

If this is skipping an unstable test:

- include info about the test failure
- submit an unstable-test issue by
[choosing](https://github.com/Esri/calcite-components/issues/new/choose)
the unstable test template and filling it out

-->