ci: disable milestone automation due to process changes (#5884)
**Related Issue:** #

## Summary

I'm disabling some of the milestone automation for now. We recently
changed the way we work with milestones, which is causing funky
behavior.

The actions were created under the assumption that the only milestones
with due dates would be for our 2 week sprints. I can change the actions
if we are deciding to change how we use milestones, but we may want to
wait for Brittney to return before making that decision, since that is
her area of expertise.

<!--

Please make sure the PR title and/or commit message adheres to the
https://www.conventionalcommits.org/en/v1.0.0/ specification.

Note: If your PR only has one commit and it is NOT semantic, you will
need to either

a. add another commit and wait for the check to update
b. proceed to squash merge, but make sure the commit message is the same
as the title.

This is because of the way GitHub handles single-commit squash merges
(see https://github.com/zeke/semantic-pull-requests/issues/17)

If this is component-related, please verify that:

- [ ] feature or fix has a corresponding test
- [ ] changes have been tested with demo page in Edge

---

If this is skipping an unstable test:

- include info about the test failure
- submit an unstable-test issue by
[choosing](https://github.com/Esri/calcite-components/issues/new/choose)
the unstable test template and filling it out

-->