docs: update setFocus method across components (#6212)
**Related Issue:** #6139

## Summary
Adds context ~to the `panel` and `popover` docs~ **across our
components** for the `setFocus` refactor. ~~Where when called will set
focus on the component's first focusable element.~~

Where when called will either set focus on the component or its first
focusable element. Some components first focusable element will be a
close button.

Unchanged components with `setFocus` where focus is set on the
component, such as `calcite-action`, and the context of "Set focus on
the component" is still valid.