fix(button): truncate long button text (#6664)
**Related Issue:** #5660

## Summary

Truncate long text on fixed-width components that should not grow in
height.

To avoid unnecessary noise, detect the `buttons` with text that has been
truncated and apply `title attribute` conditionally only to these, to
reveal the full text for sighted users. This frees up the
`calcite-tooltip` to provide additional context.

Added an e2e test to ensure the long text is truncated and the `title
attribute` is added when it is.