feat(chip-group): Add Chip Group component (#6075)
**Related Issue:**
[#1933](https://github.com/Esri/calcite-components/issues/1933)

###  ~~Opening as a draft~~ Ready for review :) 

## Summary
This still needs work but want to get a branch up so folks can preview
and provide feedback.

This PR adds the proposed `chip-group` component and some associated
changes to `chip`.
- `chip-group` can be used for spacing and keyboard navigation of
selectable or non-selectable `chip` components.
- `chip-group` defaults to `selection-mode:none`, but supports `single`,
`single-persist`, and `multiple` selection modes.

To do / looking for feedback:
- Currently, users need to add `selectable` property to child `chip`
components to enable the selection capability. This could potentially be
managed by the parent `chip-group`: conditionally applied via an
internal property based on `selection-mode`, but that creates a weird
situation where `selectable` is not documented but the `selected`
property is public, since that will always need to be available to a
user. Open to thoughts there.
  
- Currently, once dismissed, the previous chip is focused. If the first
chip is dismissed, focus is moved to the "next first chip" as it takes
the place - does this make sense?
  
  - Add more tests as needed.
  - Audit of keyboard navigation.
  - Need an a11y audit from @geospatialem :)
  - Style cleanup / design feedback from @SkyeSeitz @ashetland  :)
  - General pattern feedback :)



https://user-images.githubusercontent.com/4733155/208154126-bdc706de-8c25-4d64-b7e7-988952d96043.mov