fix(radio-button-group): no longer focus first radio button on label click and adds `setFocus` method. (#7050)
**Related Issue:**  #6698

## Summary

This PR will remove the focus flash from the first
`calcite-radio-button` in group on label click

Additional enhancements:
- Adds `setFocus( )` method in `calcite-radio-button-group`
- Users will be able to select label text in chrome. #6357 

Note:
With this PR, `delegateFocus` is removed from
`calcite-radio-button-group` which is causing #6698. Because the
tabIndex is set to 0 for the first focusable element on page load to
enable keyboard navigation into the group it makes `delegateFocus` to
apply focus styles on the first focusable element when the user clicks
on the label (invokes .focus( )) resulting in the behavior mentioned in
the issue 6698. To resolve this, `delegateFocus` is removed and added
`setFocus( )` method.